requirements/requirements.txt


pandas
openpyxl
jupyter
pyyaml